backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
haystack-ai
pgvector-haystack
ollama-haystack
sentence-transformers
python-dotenv
sqlalchemy
pydantic-settings
requests
pdfplumber
beautifulsoup4
lxml
apscheduler
